refactor: expose window globals, delete old files
