+++ packages/client/src/index.html
@@ -46,6 +46,7 @@
   </head>
   <body>
     <script type="module" src="./index.tsx"></script>
+    <script type="module" src="./exposeGlobals.ts"></script>
     <div id="react-root"></div>
   </body>
 </html>
